Completed TUI build and CLI build for day two.
The TUI now operates as designed. A 5 GB file loads in about 13 seconds through local testing. The TUI loads, you can scroll through all the results, and you can see the request and response loaded on panels on demand. Lots of work needs to be done on the visual aspect, and lots of features still need to be built, but the basic mechanics are in place now.